refactor: декомпозировал вебсокет, роут - в ws/routes, менеджер - в ws/manager, в мейне сопутствующие правки
